+++ Sources/OSS/Models/ModelObjectBasic.swift
@@ -56,6 +56,20 @@ public struct DeletedInfo: Sendable {
     public var deleteMarkerVersionId: Swift.String?
 }
 
+public struct Delete: Sendable {
+    /// Specifies whether to enable the Quiet return mode.
+    /// The DeleteMultipleObjects operation provides the following return modes: Valid value: true,false
+    public var quiet: Swift.Bool?
+
+    /// The container that stores information about you want to delete objects.
+    public var objects: [DeleteObject]?
+
+    public init(quiet: Bool? = nil, objects: [DeleteObject]? = nil) {
+        self.quiet = quiet
+        self.objects = objects
+    }
+}
+
 /// The request for the PutObject operation.
 public struct PutObjectRequest: RequestModel {
     public var commonProp: RequestModelProp
@@ -951,23 +965,17 @@ public struct DeleteMultipleObjectsRequest: RequestModel {
     /// Sees <see cref="Models.EncodingType"/> for supported values.
     public var encodingType: Swift.String?
 
-    /// Specifies whether to enable the Quiet return mode.
-    /// The DeleteMultipleObjects operation provides the following return modes: Valid value: true,false
-    public var quiet: Swift.Bool?
-
-    /// The container that stores information about you want to delete objects.
-    public var objects: [DeleteObject]?
+    /// The delete object containing quiet mode and object identifiers.
+    public var delete: Delete?
 
     public init(bucket: String? = nil,
-                objects: [DeleteObject]? = nil,
-                quiet: Bool? = nil,
+                delete: Delete? = nil,
                 encodingType: String? = nil,
                 commonProp: RequestModelProp? = nil)
     {
         self.bucket = bucket
         self.encodingType = encodingType
-        self.quiet = quiet
-        self.objects = objects
+        self.delete = delete
         self.commonProp = commonProp ?? RequestModelProp()
     }
 }
@@ -1008,6 +1016,39 @@ public struct CleanRestoredObjectRequest: RequestModel {
 /// The result for the CleanRestoredObject operation.
 public struct CleanRestoredObjectResult: ResultModel {
     public var commonProp: ResultModelProp = .init()
+}
+
+public struct SealAppendObjectRequest: RequestModel {
+    public var commonProp: RequestModelProp
+
+    /// The name of the bucket
+    public var bucket: Swift.String?
+
+    /// The name of the object.
+    public var key: Swift.String?
+
+    /// Specifies the expected length of the object when you call the SealAppendObject operation.
+    /// OSS checks whether this length matches the actual length of the object. If the lengths do not match, the request fails and the PositionNotEqualToLength error is returned.
+    public var position: Swift.Int?
+
+    public init(
+        bucket: String? = nil,
+        key: String? = nil,
+        position: Int? = nil,
+        commonProp: RequestModelProp? = nil
+    ) {
+        self.bucket = bucket
+        self.key = key
+        self.position = position
+        self.commonProp = commonProp ?? RequestModelProp()
+    }
+}
+
+public struct SealAppendObjectResult: ResultModel {
+    public var commonProp: ResultModelProp = .init()
+
+    /// The time in GMT format when the SealAppendObject operation was first performed on the object. This timestamp does not change even if the operation is performed again.
+    public var sealedTime: String? { return commonProp.headers?[caseInsensitive: "x-oss-sealed-time"] }
 }
 
 extension PutObjectResult: Sendable {}