+++ Sources/OSS/SdkInfo.swift
@@ -2,7 +2,7 @@
 public enum SdkInfo {
     // version
     static let major: String = "0"
-    static let minor: String = "2"
+    static let minor: String = "3"
     static let patch: String = "0"
     static let tag: String = ""