+++ Sources/OSS/Transform/SerdeObjectBasic.swift
@@ -402,10 +402,10 @@ extension Serde {
 
         var xmlBody = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>"
         xmlBody.append("<Delete>")
-        if let quiet = request.quiet {
+        if let quiet = request.delete?.quiet {
             xmlBody.append("<Quiet>\(quiet)</Quiet>")
         }
-        if let objects = request.objects {
+        if let objects = request.delete?.objects {
             for object in objects {
                 if let key = object.key {
                     xmlBody.append("<Object><Key>\(key.escape())</Key>")
@@ -491,6 +491,24 @@ extension Serde {
 
     static func deserializeCleanRestoredObject(
         _: inout CleanRestoredObjectResult,
+        _: inout OperationOutput
+    ) throws {}
+}
+
+// MARK: - SealAppendObject
+
+extension Serde {
+    static func serializeSealAppendObject(
+        _ request: inout SealAppendObjectRequest,
+        _ input: inout OperationInput
+    ) throws {
+        if let value = request.position {
+            input.parameters["position"] = "\(value)"
+        }
+    }
+
+    static func deserializeSealAppendObject(
+        _: inout SealAppendObjectResult,
         _: inout OperationOutput
     ) throws {}
 }