+++ Tests/OSSExtensionIntegrationTests/BaseTestCase.swift
@@ -17,9 +17,9 @@ public class BaseTestCase: XCTestCase {
     private var _client: Client? = nil
     private var _invalidClient: Client? = nil
 
-    let bucketNamePrefix = "swift-sdk-test-bucket-"
-    let objectNamePrefix = "swift-sdk-test-object-"
-    let fileNamePrefix = "swift-sdk-test-file-"
+    let bucketNamePrefix = "oss-sdk-test-swift-bucket-"
+    let objectNamePrefix = "oss-sdk-test-swift-object-"
+    let fileNamePrefix = "oss-sdk-test-swift-file-"
 
     let tempDir = NSTemporaryDirectory() + "swift-sdk-test_" + NSUUID().uuidString
     #if os(Windows)
@@ -82,7 +82,7 @@ public class BaseTestCase: XCTestCase {
         return ProcessInfo.processInfo.environment["OSS_TEST_PAYER_UID"] ?? ""
     }
 
-    var bucketName = "object-extension-test-\(Int(Date().timeIntervalSince1970))"
+    var bucketName = "oss-sdk-test-swift-bucket-\(Int(Date().timeIntervalSince1970))"
 
     var client: Client?
 
@@ -209,7 +209,12 @@ public class BaseTestCase: XCTestCase {
                 let object = DeleteObject(key: version.key, versionId: version.versionId)
                 objects.append(object)
             }
-            let deleteRequest = DeleteMultipleObjectsRequest(bucket: bucket, objects: objects)
+            let deleteRequest = DeleteMultipleObjectsRequest(
+                bucket: bucket,
+                delete: Delete(
+                    objects: objects
+                )
+            )
             try await assertNoThrow(await client.deleteMultipleObjects(deleteRequest))
         }
         if let deleteMarkers = listVersionResult.deleteMarkers {
@@ -218,7 +223,12 @@ public class BaseTestCase: XCTestCase {
                 let object = DeleteObject(key: deleteMarker.key, versionId: deleteMarker.versionId)
                 objects.append(object)
             }
-            let deleteRequest = DeleteMultipleObjectsRequest(bucket: bucket, objects: objects)
+            let deleteRequest = DeleteMultipleObjectsRequest(
+                bucket: bucket,
+                delete: Delete(
+                    objects: objects
+                )
+            )
             try await assertNoThrow(await client.deleteMultipleObjects(deleteRequest))
         }
 
@@ -232,7 +242,12 @@ public class BaseTestCase: XCTestCase {
                 let object = DeleteObject(key: content.key)
                 objects.append(object)
             }
-            let deleteRequest = DeleteMultipleObjectsRequest(bucket: bucket, objects: objects)
+            let deleteRequest = DeleteMultipleObjectsRequest(
+                bucket: bucket,
+                delete: Delete(
+                    objects: objects
+                )
+            )
             try await assertNoThrow(await client.deleteMultipleObjects(deleteRequest))
         }